+++ src/projects-v2/MealCalendar/src/play.js
@@ -1,0 +1,1 @@
+import {} from './index'
